requirements.txt
python-dotenv>=1.2.1
streamlit
requests
snowflake
pandas